Move Android studio calls out of GradleProfilerClient to a separate class
